+++ src/main/java/cech12/extendedmushrooms/init/ModItems.java
@@ -24,8 +24,8 @@ public class ModItems {
     public static final RegistryObject<Item> MUSHROOM_CHEST_BOAT = ITEMS.register("mushroom_chest_boat", () -> new MushroomBoatItem(MushroomWoodType.MUSHROOM, true));
     public static final RegistryObject<Item> GLOWSHROOM_BOAT = ITEMS.register("glowshroom_boat", () -> new MushroomBoatItem(MushroomWoodType.GLOWSHROOM, false));
     public static final RegistryObject<Item> GLOWSHROOM_CHEST_BOAT = ITEMS.register("glowshroom_chest_boat", () -> new MushroomBoatItem(MushroomWoodType.GLOWSHROOM, true));
-    public static final RegistryObject<Item> POISONOUS_MUSHROOM_BOAT = ITEMS.register("poisonous_mushroom_boat", () -> new MushroomBoatItem(MushroomWoodType.POISONOUS_MUSHROOM, false));
-    public static final RegistryObject<Item> POISONOUS_MUSHROOM_CHEST_BOAT = ITEMS.register("poisonous_mushroom_chest_boat", () -> new MushroomBoatItem(MushroomWoodType.POISONOUS_MUSHROOM, true));
+    public static final RegistryObject<Item> PARROT_WAXCAP_BOAT = ITEMS.register("poisonous_mushroom_boat", () -> new MushroomBoatItem(MushroomWoodType.PARROT_WAXCAP, false));
+    public static final RegistryObject<Item> PARROT_WAXCAP_CHEST_BOAT = ITEMS.register("poisonous_mushroom_chest_boat", () -> new MushroomBoatItem(MushroomWoodType.PARROT_WAXCAP, true));
     public static final RegistryObject<Item> HONEY_WAXCAP_BOAT = ITEMS.register("honey_fungus_boat", () -> new MushroomBoatItem(MushroomWoodType.HONEY_WAXCAP, false));
     public static final RegistryObject<Item> HONEY_WAXCAP_CHEST_BOAT = ITEMS.register("honey_fungus_chest_boat", () -> new MushroomBoatItem(MushroomWoodType.HONEY_WAXCAP, true));
 
@@ -33,10 +33,10 @@ public class ModItems {
     public static final RegistryObject<Item> MUSHROOM_HANGING_SIGN = ITEMS.register("mushroom_hanging_sign", () -> new HangingSignItem(ModBlocks.MUSHROOM_HANGING_SIGN.get(), ModBlocks.MUSHROOM_WALL_HANGING_SIGN.get(), (new Item.Properties()).stacksTo(16)));
     public static final RegistryObject<Item> GLOWSHROOM_SIGN = ITEMS.register("glowshroom_sign", () -> new SignItem((new Item.Properties()).stacksTo(16), ModBlocks.GLOWSHROOM_STANDING_SIGN.get(), ModBlocks.GLOWSHROOM_WALL_SIGN.get()));
     public static final RegistryObject<Item> GLOWSHROOM_HANGING_SIGN = ITEMS.register("glowshroom_hanging_sign", () -> new HangingSignItem(ModBlocks.GLOWSHROOM_HANGING_SIGN.get(), ModBlocks.GLOWSHROOM_WALL_HANGING_SIGN.get(), (new Item.Properties()).stacksTo(16)));
-    public static final RegistryObject<Item> POISONOUS_MUSHROOM_SIGN = ITEMS.register("poisonous_mushroom_sign", () -> new SignItem((new Item.Properties()).stacksTo(16), ModBlocks.POISONOUS_MUSHROOM_STANDING_SIGN.get(), ModBlocks.POISONOUS_MUSHROOM_WALL_SIGN.get()));
-    public static final RegistryObject<Item> POISONOUS_MUSHROOM_HANGING_SIGN = ITEMS.register("poisonous_mushroom_hanging_sign", () -> new HangingSignItem(ModBlocks.POISONOUS_MUSHROOM_HANGING_SIGN.get(), ModBlocks.POISONOUS_MUSHROOM_WALL_HANGING_SIGN.get(), (new Item.Properties()).stacksTo(16)));
-    public static final RegistryObject<Item> HONEY_FUNGUS_SIGN = ITEMS.register("honey_fungus_sign", () -> new SignItem((new Item.Properties()).stacksTo(16), ModBlocks.HONEY_WAXCAP_STANDING_SIGN.get(), ModBlocks.HONEY_WAXCAP_WALL_SIGN.get()));
-    public static final RegistryObject<Item> HONEY_FUNGUS_HANGING_SIGN = ITEMS.register("honey_fungus_hanging_sign", () -> new HangingSignItem(ModBlocks.HONEY_WAXCAP_HANGING_SIGN.get(), ModBlocks.HONEY_WAXCAP_WALL_HANGING_SIGN.get(), (new Item.Properties()).stacksTo(16)));
+    public static final RegistryObject<Item> PARROT_WAXCAP_SIGN = ITEMS.register("poisonous_mushroom_sign", () -> new SignItem((new Item.Properties()).stacksTo(16), ModBlocks.PARROT_WAXCAP_STANDING_SIGN.get(), ModBlocks.PARROT_WAXCAP_WALL_SIGN.get()));
+    public static final RegistryObject<Item> PARROT_WAXCAP_HANGING_SIGN = ITEMS.register("poisonous_mushroom_hanging_sign", () -> new HangingSignItem(ModBlocks.PARROT_WAXCAP_HANGING_SIGN.get(), ModBlocks.PARROT_WAXCAP_WALL_HANGING_SIGN.get(), (new Item.Properties()).stacksTo(16)));
+    public static final RegistryObject<Item> HONEY_WAXCAP_SIGN = ITEMS.register("honey_fungus_sign", () -> new SignItem((new Item.Properties()).stacksTo(16), ModBlocks.HONEY_WAXCAP_STANDING_SIGN.get(), ModBlocks.HONEY_WAXCAP_WALL_SIGN.get()));
+    public static final RegistryObject<Item> HONEY_WAXCAP_HANGING_SIGN = ITEMS.register("honey_fungus_hanging_sign", () -> new HangingSignItem(ModBlocks.HONEY_WAXCAP_HANGING_SIGN.get(), ModBlocks.HONEY_WAXCAP_WALL_HANGING_SIGN.get(), (new Item.Properties()).stacksTo(16)));
 
     public static final RegistryObject<Item> GRILLED_MUSHROOM = ITEMS.register("grilled_mushroom", () -> new Item((new Item.Properties()).food((new FoodProperties.Builder()).nutrition(2).saturationMod(0.3F).build())));
     public static final RegistryObject<Item> MUSHROOM_BREAD = ITEMS.register("mushroom_bread", () -> new Item((new Item.Properties()).food((new FoodProperties.Builder()).nutrition(5).saturationMod(0.3F).build())));